+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Every entry has a category for which we use the following visual abbreviations:
 
 ## Unreleased
 
+- 🐞 The `vast status` command does not collect status information from sources
+  and sinks any longer. They were often too busy to respond, leading to a long
+  delay before the command completed.
+  [#1234](https://github.com/tenzir/vast/pull/1234)
+
 - 🎁 A new tracepoint `meta_index_lookup(us, candidates)` can be used, that
   is called whenever a meta index lookup is completed. It gets two arguments,
   the number of microseconds spent resolving the lookup and the number of